Step3 질문 삭제하기 리팩터링 (#154)
* refactor: question 에게 validateOwner를 시키도록 수정

* refactor: delete() 메서들르 추가하여 setter를 제거하도록 수정

* docs: 요구사항 명세서 추가

* refactor: question delete하고 history를 추가하는 부분을 Question의 책임으로 이동

* refactor: 게시판을 삭제하고 히스토리를 추가하는 것에 대한 기능 테스트코드 추가

* refactor: QnaService에 작성된 delete와 history를 save하는 로직을 Question으로 이동 및 해당 로직에 대한 테스트 코드 추가

* refactor: 구현한 기능에 대한 체크리스트 추가

* refactor: validateOwner로 검증하고 deleteHistories를 생성하도록 수정

* refactor: 불필요한 선언 제거 및 불변성을 유지하기 위해 List.of사용 및 리턴하도록 함수 수정 및 테스트 코드 수정

+++ src/main/java/qna/domain/Question.java
@@ -3,6 +3,7 @@ package qna.domain;
 import qna.CannotDeleteException;
 
 import javax.persistence.*;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
@@ -42,27 +43,46 @@ public class Question extends AbstractEntity {
         return this;
     }
 
-    public void deleteAllAnswersAndAddHistories(List<DeleteHistory> deleteHistories) {
-        answers.deleteAllAndAddHistories(deleteHistories);
+    public List<DeleteHistory> deleteAllAndAddHistories(User loginUser) throws CannotDeleteException {
+        validateOwner(loginUser);
+
+        List<DeleteHistory> deleteHistories = deleteAndAddHistory();
+
+        return answers.deleteAllAndAddHistories(deleteHistories);
     }
 
-    public void validateAnswerOwner(User loginUser) throws CannotDeleteException {
+    private void validateOwner(User loginUser) throws CannotDeleteException {
+        validateIsOwner(loginUser);
+        validateAnswerOwner(loginUser);
+    }
+
+    private void validateAnswerOwner(User loginUser) throws CannotDeleteException {
         answers.validateOwners(loginUser);
     }
 
-    public void validateIsOwner(User loginUser) throws CannotDeleteException {
+    private void validateIsOwner(User loginUser) throws CannotDeleteException {
         if (!isOwner(loginUser)) {
             throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
         }
     }
 
-    public boolean isOwner(User writer) {
+    private boolean isOwner(User writer) {
         return this.writer.equals(writer);
+    }
+
+    private List<DeleteHistory> deleteAndAddHistory() {
+        delete();
+        return List.of(new DeleteHistory(ContentType.QUESTION, id, writer, LocalDateTime.now()));
     }
 
     public void addAnswer(Answer answer) {
         answer.toQuestion(this);
         answers.add(answer);
+    }
+
+
+    public boolean isDeleted() {
+        return deleted;
     }
 
     public Long getId() {
@@ -73,12 +93,8 @@ public class Question extends AbstractEntity {
         return writer;
     }
 
-    public boolean isDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        this.deleted = deleted;
+    public void delete() {
+        this.deleted = true;
     }
 
     @Override
